L5 V3 C2/2

+++ lab5/duplicate.sh
@@ -7,10 +7,16 @@ for i in $(tree -fi $direct | head -n-2); do
 	then
 		echo -e "DIRECTORY:\e[33m$i\e[0m";
 	else
+		[ ! -L $i ] && [ ! -f $i ] && continue;
 		echo -e "FILE:\e[36m$i\e[0m";
+		if [ -L $i ]; then
+			echo -e "\e[35mLINK: $i\e[0m"
+			continue
+		fi;
 		sha=$(sha256sum $i)
 		shaOnly=$(cut -d" " -f1 <<< "$sha")
 		file=$(cut -d" " -f3 <<< "$(grep $shaOnly /tmp/dupl)")
+		[ "$file" == "$i" ] && continue;
 		[ -z "$file" ] && echo -e "\e[32mNEW: $i\e[0m" || echo -e "\e[34mCOPY: $file\e[0m"
 		if [ -n "$file" ]; then
 			rm $i
